# Conflicts:
#	src/Interpreters/ConvertFunctionOrLikeVisitor.cpp
#	tests/queries/0_stateless/02226_analyzer_or_like_combine.oldanalyzer.reference

+++ src/AggregateFunctions/CrossTab.h
@@ -135,13 +135,13 @@ struct CrossTabData
         Float64 sum = 0.0;
         for (const auto & [key, value_ab_uint] : count_ab)
         {
-            const Float64 value_a = count_a.at(key.items[UInt128::_impl::little(0)]);
-            const Float64 value_b = count_b.at(key.items[UInt128::_impl::little(1)]);
+            const Float64 value_a = static_cast<Float64>(count_a.at(key.items[UInt128::_impl::little(0)]));
+            const Float64 value_b = static_cast<Float64>(count_b.at(key.items[UInt128::_impl::little(1)]));
 
             assert(value_a > 0 && "frequency of value `a` must be positive");
             assert(value_b > 0 && "frequency of value `b` must be positive");
 
-            const Float64 value_ab = value_ab_uint;
+            const Float64 value_ab = static_cast<Float64>(value_ab_uint);
 
             sum += (value_ab / value_a) * (value_ab / value_b);
         }
